requirements.txt
aiohttp>=3.9.0
aiofiles>=23.2.1
redis[asyncio]>=5.0.0
pydantic>=2.0.0
pytest-asyncio>=0.23.0
